Teach worker threads how to pause
Inspector uses a V8 interrupt and a nested event loop to pause
workers. Extract this functionality out of the worker thread debugger and
add it to the worker thread itself. Call these pause/resume functions for
dedicated workers when a frame freeze's itself as well.

Test this new functionality in unit tests.

BUG=907125

Change-Id: I9cd2c27803ecf30ad8e6ac780b05a055234e3ffb